+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add RKDE model implementation (https://github.com/openvinotoolkit/anomalib/pull/821)
+- Add Visual Anomaly (VisA) dataset adapter (<https://github.com/openvinotoolkit/anomalib/pull/824>)
 - Add Synthetic anomalous dataset for validation and testing (https://github.com/openvinotoolkit/anomalib/pull/822)
 - Add Detection task type support (https://github.com/openvinotoolkit/anomalib/pull/822)
 - Add UCSDped and Avenue dataset implementation (https://github.com/openvinotoolkit/anomalib/pull/822)